add index routes/models/controllers
