+++ controllers/budgets.js
@@ -1,0 +1,9 @@
+import { Budget } from '../models/budgets.js'
+
+function index(req, res) {
+
+}
+
+export {
+  index
+}
