+++ controllers/expenses.js
@@ -1,0 +1,9 @@
+import { Expense } from '../models/expenses.js'
+
+function index(req, res) {
+  
+}
+
+export {
+  index
+}
