+++ controllers/goals.js
@@ -1,0 +1,10 @@
+import { Goal } from '../models/goal.js'
+
+function index(req, res) {
+
+}
+
+export {
+  index
+}
+
