+++ controllers/incomes.js
@@ -1,0 +1,9 @@
+import { Income } from '../models/incomes.js'
+
+function index(req, res) {
+
+}
+
+export {
+  index
+}
