+++ routes/budgets.js
@@ -1,16 +1,16 @@
 import { Router } from 'express'
-import * as profilesCtrl from '../controllers/profiles.js'
+import * as budgetsCtrl from '../controllers/budgets.js'
 import { decodeUserFromToken, checkAuth } from '../middleware/auth.js'
 
 const router = Router()
 
 /*---------- Public Routes ----------*/
 
-
 /*---------- Protected Routes ----------*/
 router.use(decodeUserFromToken)
 
-router.get('/', checkAuth, profilesCtrl.index)
+router.get('/', checkAuth, budgetsCtrl.index)
 
-
-export { router }
+export{
+  router
+}
