+++ routes/expenses.js
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import * as profilesCtrl from '../controllers/profiles.js'
+import * as expensesCtrl from '../controllers/expenses.js'
 import { decodeUserFromToken, checkAuth } from '../middleware/auth.js'
 
 const router = Router()
@@ -10,7 +10,8 @@ const router = Router()
 /*---------- Protected Routes ----------*/
 router.use(decodeUserFromToken)
 
-router.get('/', checkAuth, profilesCtrl.index)
+router.get('/', checkAuth, expensesCtrl.index)
 
-
-export { router }
+export {
+  router
+}
