+++ routes/goals.js
@@ -1,16 +1,16 @@
 import { Router } from 'express'
-import * as profilesCtrl from '../controllers/profiles.js'
+import * as goalsCtrl from  '../controllers/goals.js'
 import { decodeUserFromToken, checkAuth } from '../middleware/auth.js'
 
 const router = Router()
 
 /*---------- Public Routes ----------*/
 
-
 /*---------- Protected Routes ----------*/
 router.use(decodeUserFromToken)
 
-router.get('/', checkAuth, profilesCtrl.index)
+router.get('/', checkAuth, goalsCtrl.index)
 
-
-export { router }
+export {
+  router
+}
