+++ server.js
@@ -5,6 +5,10 @@ import cors from 'cors'
 
 import { router as profilesRouter } from './routes/profiles.js'
 import { router as authRouter } from './routes/auth.js'
+import { router as budgetsRouter} from './routes/budgets.js'
+import { router as expensesRouter } from './routes/expenses.js'
+import { router as goalsRouter } from './routes/goals.js'
+import { router as incomesRouter } from './routes/incomes.js'
 
 import('./config/database.js')
 
@@ -16,6 +20,10 @@ app.use(express.json())
 
 app.use('/api/profiles', profilesRouter)
 app.use('/api/auth', authRouter)
+app.use('/api/budgets', budgetsRouter)
+app.use('/api/expenses', expensesRouter)
+app.use('/api/goals', goalsRouter)
+app.use('/api/incomes', incomesRouter)
 
 app.use(function (req, res, next) {
   res.status(404).json({ err: "Not found" })
